Create LogStuff mixin and DRY out code
Shared in several places and be better shared
app-wide in any event.
